third_party/WebKit/LayoutTests/fast/xmlhttprequest/xmlhttprequest-withcredentials-before-open.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
    <script>
        description('This tests that the XMLHttpRequest withCredentials attribute is modifiable prior to being in the OPENED state.');
        var xhr;

        xhr = new XMLHttpRequest();
        evalAndLog("xhr.withCredentials = true;");
        evalAndLog("xhr.open('GET', 'http://mydomain');");
        shouldBeTrue('xhr.withCredentials');
    </script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
