third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-withcredentials-before-open-expected.txt

This tests that the XMLHttpRequest withCredentials attribute is modifiable prior to being in the OPENED state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

xhr.withCredentials = true;
xhr.open('GET', 'http://mydomain');
PASS xhr.withCredentials is true
PASS successfullyParsed is true

TEST COMPLETE
